Set USER INFO controller

+++ src/main/java/io/bank/management/entity/UserInfo.java
@@ -19,6 +19,6 @@ public class UserInfo {
     private String lastName;
     private String phone;
     private String adresse;
-    private int userID;
+    private Long userID;
 
 }
